+++ data-contract/src/storage_key.rs
@@ -11,4 +11,5 @@ pub enum StorageKey {
     Enumeration,
     Approval,
     Ownership,
+    Rbac,
 }
